+++ src/components/modal/index.jsx
@@ -1,64 +1,64 @@
 import React, { PropTypes } from "react";
 import Modal from "react-modal";
 import radium, { Style } from "radium";
-import { color, media, timing, zIndex } from "../../../settings.json";
+import colors from "../../styles/colors";
+import mq from "../../styles/mq";
+import timing from "../../styles/timing";
+import zIndex from "../../styles/zIndex";
 import Heading from "../heading";
-import { rgb } from "../../utils/color";
+import { rgba } from "../../utils/color";
 import propTypes from "../../utils/propTypes";
 
-const largeMQ = `(min-width: ${media.min["768"]})`;
-const modalPadding = 56;
+const largeMQ = `(min-width: ${mq.min["768"]})`;
+const modalPadding = 24;
 
 const styles = {
   overlay: {
-    backgroundColor: `rgba(${rgb(color.black)}, .4)`,
+    backgroundColor: rgba(colors.bgOverlay, 0.3),
     overflow: "hidden",
-    zIndex: zIndex.modal - 2,
+    zIndex: zIndex.modal,
   },
+
   header: {
-    borderBottom: `1px solid ${color.gray}`,
-    paddingBottom: "16px",
-    paddingTop: "16px",
+    height: "56px",
     position: "relative",
     textAlign: "center",
-    textTransform: "uppercase",
+
     [`@media ${largeMQ}`]: {
-      paddingLeft: `${modalPadding}px`,
-      paddingRight: `${modalPadding}px`,
-      paddingTop: `${modalPadding}px`,
-      paddingBottom: 0,
-      borderBottom: 0,
-      textAlign: "left",
+      height: "72px",
     },
   },
+
   contentContainer: {
-    paddingLeft: `${modalPadding}px`,
-    paddingRight: `${modalPadding}px`,
-    paddingTop: "32px",
-    paddingBottom: `${modalPadding * 2}px`,
+    padding: "40px",
+
+    [`@media ${largeMQ}`]: {
+      padding: "8px 56px",
+    },
   },
+
   actionItem: {
     position: "absolute",
     backgroundColor: "transparent",
-    top: "12px",
+    top: 0,
+    padding: "16px",
+
     [`@media ${largeMQ}`]: {
-      top: `${modalPadding}px`,
+      padding: `${modalPadding}px`,
     },
   },
+
   rightAction: {
-    right: "16px",
-    [`@media ${largeMQ}`]: {
-      right: `${modalPadding}px`,
-    },
+    right: 0,
   },
+
   leftAction: {
-    left: "16px",
-    [`@media ${largeMQ}`]: {
-      left: `${modalPadding}px`,
-    },
+    left: 0,
   },
+
   desktopTitle: {
     display: "none",
+
     [`@media ${largeMQ}`]: {
       display: "block",
       textAlign: "center",
@@ -66,9 +66,12 @@ const styles = {
       paddingBottom: "104px",
     },
   },
+
   mobileTitle: {
-    display: "block",
+    display: "inline-block",
     minHeight: "10px",
+    paddingTop: "23px",
+
     [`@media ${largeMQ}`]: {
       display: "none",
     },
@@ -80,6 +83,7 @@ function ModalComponent({
   closeModal,
   closeTimeoutMS,
   contentLabel,
+  desktopMaxHeight,
   desktopWidth,
   leftAction,
   leftActionContent,
@@ -90,50 +94,56 @@ function ModalComponent({
   style,
 }) {
   const rules = {
+    ".ReactModal__Content": {
+      outline: 0,
+    },
+
     ".ReactModal__Content--after-open.ReactModal__Content--before-close": {
       opacity: 0,
       transform: "translateY(55%)",
       transition: `opacity ${timing.default},
         transform ${timing.default}`,
     },
+
     ".ReactModal__Content--after-open": {
-      opacity: "1 !important",
-      transform: "tranlateY(0) !important",
+      opacity: 1,
+      transform: "translateY(0)",
       transition: `opacity ${timing.default},
         transform ${timing.default}`,
     },
+
     ".ReactModal__Overlay--after-open.ReactModal__Overlay--before-close": {
       opacity: 0,
-      transition: `opacity ${timing.default},
-        transform ${timing.default}`,
+      transition: `opacity ${timing.default}`,
     },
+
     ".ReactModal__Overlay--after-open": {
       opacity: 1,
-      transition: `opacity ${timing.default},
-        transform ${timing.default}`,
+      transition: `opacity ${timing.default}`,
     },
+
     ".ModalBase": {
-      background: color.white,
+      background: colors.bgPrimary,
       position: "absolute",
       overflow: "auto",
       WebkitOverflowScrolling: "touch",
       border: 0,
       borderRadius: 0,
-      boxShadow: `0 27px 50px rgba(${rgb(color.black)}, .36)`,
+      boxShadow: `0 27px 50px ${rgba(colors.bgOverlay, 0.36)}`,
       top: 0,
       left: 0,
       right: 0,
-      zIndex: zIndex.modal,
       marginLeft: "auto",
       marginRight: "auto",
       bottom: "auto",
       height: "100vh",
       width: "100%",
     },
+
     mediaQueries: {
       [largeMQ]: {
         ".ModalBase": {
-          maxHeight: "85vh",
+          maxHeight: desktopMaxHeight,
           top: "50%",
           width: desktopWidth,
           maxWidth: "1290px",
@@ -142,6 +152,7 @@ function ModalComponent({
       },
     },
   };
+
   return (
     <Modal
       isOpen={isOpen}
@@ -155,9 +166,17 @@ function ModalComponent({
         scopeSelector=".ReactModalPortal"
         rules={rules}
       />
+
       <header
         className="Modal-header clearfix"
-        style={styles.header}
+        style={[
+          styles.header,
+          title && {
+            [`@media (max-width: ${mq.max["768"]})`]: {
+              borderBottom: `1px solid ${colors.borderPrimary}`,
+            },
+          },
+        ]}
       >
         {leftAction &&
           <button
@@ -167,15 +186,18 @@ function ModalComponent({
             {leftActionContent}
           </button>
         }
-        <Heading
-          level={4}
-          size="small"
-          weight="thick"
-          override={styles.mobileTitle}
-          caps
-        >
-          {title}
-        </Heading>
+
+        {title &&
+          <Heading
+            level={4}
+            size="small"
+            weight="thick"
+            override={styles.mobileTitle}
+            caps
+          >
+            {title}
+          </Heading>
+        }
 
         {rightAction &&
           <button
@@ -201,6 +223,7 @@ function ModalComponent({
             {title}
           </Heading>
         }
+
         {children}
       </div>
     </Modal>
@@ -209,6 +232,10 @@ function ModalComponent({
 
 ModalComponent.propTypes = {
   isOpen: PropTypes.bool.isRequired,
+  children: PropTypes.oneOfType([
+    PropTypes.node,
+    PropTypes.string,
+  ]).isRequired,
   closeModal: PropTypes.func,
   leftAction: PropTypes.func,
   leftActionContent: PropTypes.node,
@@ -216,22 +243,17 @@ ModalComponent.propTypes = {
   rightActionContent: PropTypes.node,
   closeTimeoutMS: PropTypes.number,
   contentLabel: PropTypes.string,
+  desktopMaxHeight: PropTypes.string,
   desktopWidth: PropTypes.string,
-  title: PropTypes.string.isRequired,
+  title: PropTypes.string,
   style: propTypes.style,
-  children: PropTypes.oneOfType([
-    PropTypes.node,
-    PropTypes.string,
-  ]).isRequired,
 };
 
 ModalComponent.defaultProps = {
   isOpen: false,
-  title: "Modal",
+  desktopMaxHeight: "85vh",
   desktopWidth: "85%",
-  contentLabel: "Modal",
-  closeTimeoutMS: 500,
-  children: null,
+  closeTimeoutMS: timing.default,
 };
 
 ModalComponent.styles = styles;