+++ src/styles/colors.js
@@ -4,6 +4,7 @@ import { rgba } from "../utils/color";
 const bgPrimary = "#fff";
 const bgSecondary = "#f4fbfe";
 const bgOverlay = "#000";
+const bgDark = rgba("#000", 0.88);
 
 // Link colors
 const linkPrimary = "#297cbb";
@@ -49,6 +50,7 @@ export default {
   bgPrimary,
   bgSecondary,
   bgOverlay,
+  bgDark,
 
   linkPrimary,
   linkPrimaryHover,